cardio-backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
pydantic[email]==2.6.1
beanie==1.25.0
motor==3.3.2
bcrypt==4.1.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
